Combine_Files.py
This file will take all our data files and combine them into one file. If a file is "corrupted" (not in the correct format), it will not add its data to the file. At the end, a report is run to list all corrupted files' names and the percentage of error files.