Added Trapsat and Rockblock
These are heavily modified files, I added a lot to the two files, but more importantly I attempted to create a library for the functions, so we can write our main function in a separate file and call these files with the include. These are untested and some of the functionality might not work well so I left the orignals incase they do not work.
Eventually once I test these we can rename them.